+++ packages/interweave-emoji-picker/CHANGELOG.md
@@ -1,4 +1,9 @@
 # 1.3.0
+#### 🚀 New
+* Added a new `disableGroups` prop to `Picker`, which will hide the tabs and disable group headers.
+  * Emojis will now be grouped into a "none" group.
+  * Commonly used emojis and searching are still available in this mode.
+
 #### 🛠 Internal
 * Commonly used emojis now store the `hexcode` instead of the `unicode` character.